requirements.txt
amqp==2.2.2
billiard==3.5.0.3
bleach==2.1.3
celery==4.1.0
Django==2.1
django-celery-results==1.0.1
django-haystack==2.8.1
django-redis==4.9.0
eventlet==0.22.1
greenlet==0.4.13
html5lib==1.0.1
jieba==0.39
kombu==4.1.0
Markdown==2.6.11
mdx-bleach==0.1.2
mistune==0.8.4
mysqlclient==1.4.2.post1
Pillow==5.0.0
Pygments==2.3.1
PyMySQL==0.8.0
pytz==2018.3
redis==2.10.6
six==1.11.0
vine==1.1.4
webencodings==0.5.1
Whoosh==2.7.4
